Added documentation to some functions, added helper functions in functions folder. Will be adding more documentation in future just needed to commit as program is at an OK working standpoint. CURRENT ISSUE:
Original model is a resnet34 model I trained on the caltech
101 images of 101 categories of images. Current issue is caused
when the program reaches predict_on_new_images, the model has an out
of bounds error and attempts to output a category from the oringinal
model. I belive the error resides in DataModelMaker and has
something to do with how the model is being altered.